+++ api_types.ts
@@ -228,9 +228,13 @@ export interface ParamDef<T extends UnionType> {
   // and we'll wrap this into an autocomplete formula on their behalf.
   autocomplete?: MetadataFormula;
   /**
-   * The default value to be used for this parameter if it is not specified by the user.
+   * @deprecated This will be removed in a future version of the SDK. Use {@link suggestedValue} instead.
    */
-  defaultValue?: DefaultValueType<T>;
+  defaultValue?: SuggestedValueType<T>;
+  /**
+   * The suggested value to be prepopulated for this parameter if it is not specified by the user.
+   */
+  suggestedValue?: SuggestedValueType<T>;
 }
 
 /** @hidden */
@@ -261,9 +265,9 @@ export type ParamValues<ParamDefsT extends ParamDefs> = {
 } & any[]; // NOTE(oleg): we need this to avoid "must have a '[Symbol.iterator]()' method that returns an iterator."
 
 /**
- * The type of values that are allowable to be used as a {@link defaultValue} for a parameter.
+ * The type of values that are allowable to be used as a {@link suggestedValue} for a parameter.
  */
-export type DefaultValueType<T extends UnionType> = T extends ArrayType<Type.date>
+export type SuggestedValueType<T extends UnionType> = T extends ArrayType<Type.date>
   ? TypeOfMap<T> | PrecannedDateRange
   : TypeOfMap<T>;
 
@@ -629,7 +633,7 @@ export interface SyncExecutionContext extends ExecutionContext {
 // A mapping exists in coda that allows these to show up in the UI.
 // If adding new values here, add them to that mapping and vice versa.
 /**
- * Special "live" date range values that can be used as the {@link defaultValue}
+ * Special "live" date range values that can be used as the {@link suggestedValue}
  * for a date array parameter.
  *
  * Date array parameters are meant to represent date ranges. A date range can