+++ documentation/examples/column-format/reverse.ts
@@ -29,7 +29,7 @@ pack.addFormula({
       type: coda.ParameterType.Boolean,
       name: "byWord",
       description: "Reverse the text word-by-word.",
-      defaultValue: false,
+      suggestedValue: false,
       optional: true,
     }),
   ],